feat: consolidate RMS alerts into app router
